refactor(ntc-vault-cli): separate code into two crates: core, cli
