--- clients/ntc-vault-cli/src/core/mod.rs
@@ -1,4 +1,0 @@
-//! Core functionality.
-
-pub(crate) mod crypto;
-pub(crate) mod identity;
