--- clients/ntc-vault-cli/src/lib.rs
@@ -1,4 +1,0 @@
-//! Nautilus Trusted Compute Vault CLI
-
-pub mod cli;
-mod core;
