+++ ntc-vault-rust/Cargo.toml
@@ -1,0 +1,3 @@
+[workspace]
+
+members = ['ntc-vault-cli', 'ntc-vault-core']
