+++ ntc-vault-rust/ntc-vault-cli/Cargo.toml
@@ -11,10 +11,10 @@ base64 = "0.13"
 clap = { version = "3.1", features = ["derive"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_with = { version = "1.12", features = ["base64"] }
-thiserror = "1.0"
 
 # XXX: Waiting for release 0.5.0: <https://github.com/rust-cli/confy/issues/41>
 confy = { git = "https://github.com/rust-cli/confy", branch = "master" }
 
-rand = "0.8"
 rusty-sodalite = { git = "https://github.com/PiDelport/rusty-sodalite", branch = "initial-version" }
+
+ntc-vault-core = { path = "../ntc-vault-core" }
