+++ ntc-vault-rust/ntc-vault-cli/src/bin/ntc-vault.rs
@@ -1,5 +1,5 @@
 //! `ntc-vault` binary entrypoint.
 
 fn main() -> anyhow::Result<()> {
-    ntc_vault_cli::cli::commands::main()
+    ntc_vault_cli::commands::main()
 }
