+++ ntc-vault-rust/ntc-vault-cli/src/lib.rs
