+++ ntc-vault-rust/ntc-vault-core/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "ntc-vault-core"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+rand = "0.8"
+
+rusty-sodalite = { git = "https://github.com/PiDelport/rusty-sodalite", branch = "initial-version" }
