+++ ntc-vault-rust/ntc-vault-core/src/crypto.rs
@@ -4,7 +4,7 @@ use rand::{thread_rng, RngCore};
 use rusty_sodalite::types::SecureSeed;
 
 /// Generate a new secure seed using [`thread_rng`].
-pub(crate) fn generate_secure_seed() -> Result<SecureSeed, rand::Error> {
+pub fn generate_secure_seed() -> Result<SecureSeed, rand::Error> {
     let mut seed = SecureSeed::default();
     thread_rng().try_fill_bytes(&mut seed)?;
     Ok(seed)
