+++ ntc-vault-rust/ntc-vault-core/src/lib.rs
@@ -1,0 +1,4 @@
+//! Core functionality.
+
+pub mod crypto;
+pub mod identity;
